reorganize tests for mars-utils
